updated epic for catalogItems to make it consider categoryId, added Catalog page to the routes
