+++ src/components/Top/Top.jsx
@@ -1,14 +1,30 @@
+import { useEffect } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { topSalesRequest } from "../../store/slices";
+import Loader from "../Loader/Loader";
 import Card from "../Card/Card";
 import "./Top.css";
 
-const Top = ({items}) => {
+const Top = () => {
+    const { topSales, topSalesLoading, topSalesError } = useSelector(state => state.topSales)
+    const dispatch = useDispatch();
+
+    useEffect(() => {
+        dispatch(topSalesRequest())
+    }, [])
     return (
-        <section className="top-sales">
-            <h2 className="text-center">Хиты продаж!</h2>
-            <div className="row">
-                {items.map(item => <Card key={item.id} item={item} />)}
-            </div>
-          </section>
+        <>
+            {topSalesLoading && <Loader />}
+            {
+                topSalesError ? <div>Error</div> : 
+                <section className="top-sales">
+                    <h2 className="text-center">Хиты продаж!</h2>
+                    <div className="row">
+                        {topSales.map(item => <Card key={item.id} item={item} />)}
+                    </div>
+                </section>
+            }
+        </>
     )
 }
