+++ src/store/epics.js
@@ -2,7 +2,7 @@ import { ofType } from "redux-observable";
 import { map, switchMap, catchError } from "rxjs/operators";
 import { ajax } from "rxjs/ajax";
 import { of } from "rxjs";
-import { catalogItemsFailure, catalogItemsRequest, catalogItemsSuccess, topSalesFailure, topSalesRequest, topSalesSuccess } from "./slices";
+import { catalogItemsFailure, catalogItemsRequest, catalogItemsSuccess, categoryIdSelect, topSalesFailure, topSalesRequest, topSalesSuccess } from "./slices";
 
 const url = process.env.REACT_APP_BASE_URL;
 
@@ -15,11 +15,15 @@ export const topSalesEpic = action$ => action$.pipe(
     )
 );
 
-export const catalogItemsEpic = action$ => action$.pipe(
-    ofType(catalogItemsRequest.type),
-    switchMap(() => ajax.getJSON(`${url}/items`).pipe(
-            map(res => catalogItemsSuccess(res)),
-            catchError(error => of(catalogItemsFailure(error)))
-        )
+export const catalogItemsEpic = (action$, state$) => action$.pipe(
+    ofType(catalogItemsRequest.type, categoryIdSelect.type),
+    switchMap(() => {
+            const categoryId = state$.value.catalogItems.categoryId;
+            const endpoint = categoryId ? `${url}/items?categoryId=${categoryId}` : `${url}/items`;
+            return ajax.getJSON(endpoint).pipe(
+                map(res => catalogItemsSuccess(res)),
+                catchError(error => of(catalogItemsFailure(error)))
+            )
+        }
     )
 );
